Generate Autocorrelated Noise

Generates temporally autocorrelated random numbers with a mean, standard deviation, and autocorrelation you specify.

colored_noise(timesteps, mean, sd, phi)

Arguments

timesteps
The number of temporally autocorrelated random numbers (one per timestep) you want.

mean
The mean of the temporally autocorrelated random numbers.

sd
The standard deviation of the temporally autocorrelated random numbers.

phi
The temporal autocorrelation. 0 is white noise (uncorrelated), positive values are red noise (directly correlated) and negative values are blue noise (inversely correlated).

Value

A vector of temporally autocorrelated random numbers.

Examples

rednoise <- colored_noise(timesteps = 30, mean = 0.5, sd = 0.2, phi = 0.3)
rednoise

#>  [1] 0.37021023 0.41566726 0.80294995 0.99961101 0.48412124 0.42995334
#>  [7] 0.43914767 0.48285556 0.32007978 0.65416608 0.74251701 0.67297348
#> [13] 0.75580575 0.42160936 0.09211461 0.35413028 0.31183468 0.63184455
#> [19] 0.39399504 0.57217549 0.38275878 0.44158507 0.34031239 0.29727835
#> [25] 0.23046872 0.23543332 0.59466270 0.48485054 0.14622012 0.44496900