Generate Multiple Cross-Correlated & Autocorrelated Variables

Generates random variables that are correlated to each other and temporally autocorrelated.

colored_multi_rnorm(timesteps, mean, sd, phi, covMatrix)

Arguments

timesteps
The number of temporally autocorrelated random numbers (one per timestep) you want.

mean
A vector giving the mean of each variable.

sd
A vector giving the standard deviation of each variable.

phi
A vector giving the temporal autocorrelation of each variable.

covMatrix
A valid covariance matrix. The number of rows/columns must match the length of the mu, sigma, and phi vectors.

Value

A matrix with as many rows as timesteps and as many columns as mu/sigma/phi values.

Examples

cov <- matrix(c(0.037, 0.044, -0.048, 0.044, 0.247, -0.008, -0.047, -0.008, 0.074), nrow = 3)
test <- colored_multi_rnorm(100, c(0, 3, 5), c(1, 0.5, 1), c(0.5, -0.3, 0), cov)
var(test)

#>             [,1]        [,2]         [,3]
#> [1,]  0.04857169 0.026601969 -0.039341143
#> [2,]  0.02660197 0.277903438  0.008478241
#> [3,] -0.03934114 0.008478241  0.063574516

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

test %>% as.data.frame() %>% summarize_all(.funs = c("mean", "sd", "autocorrelation"))

#>      V1_mean  V2_mean  V3_mean     V1_sd     V2_sd     V3_sd V1_autocorrelation
#> 1 -0.0287741 2.957824 5.004534 0.2203899 0.5271655 0.2521399          0.5234464
#>   V2_autocorrelation V3_autocorrelation
#> 1         -0.3132854        -0.02673646